# refactor(experimental): use DrainOuterGeneric helper on codec type mappings