@lorisleiva 2024-03-15
@lithdew Would you have any recommendation on how I could reproduce your performance issues? I copy/pasted the code you shared in the original issue for the benchmark above but it didn't make any difference.