@lithdew 2024-03-15
Would you be able to try a trace with this?

```ts
import {
  getDataEnumCodec,
  getStructCodec,
  getStringCodec,
  getU64Codec,
  getU16Codec,
  getNullableCodec,
  getArrayCodec,
  getUnitCodec,
  getBase58Codec,
  getBooleanCodec,
  getScalarEnumCodec,
  getU8Codec,
  type Codec,
} from "@solana/web3.js";

export enum Currency {
  sol = "sol",
  like = "like",
  usdc = "usdc",
}

export const CURRENCY_CODEC = getScalarEnumCodec(Currency);

export const PUBLIC_KEY = getStringCodec({
  size: 32,
  encoding: getBase58Codec(),
});

export const CREATOR = getStructCodec([
  ["address", PUBLIC_KEY],
  ["verified", getBooleanCodec()],
  ["share", getU8Codec()],
]);

export const SPLIT = getStructCodec([
  ["address", PUBLIC_KEY],
  ["share", getU8Codec()],
]);

export const PASS_INSTRUCTION_DATA = getDataEnumCodec([
  [
    "create_pass",
    getStructCodec([
      ["name", getStringCodec()],
      ["symbol", getStringCodec()],
      ["uri", getStringCodec()],
      ["price", getU64Codec()],
      ["token", CURRENCY_CODEC],
      ["supply", getU64Codec()],
      ["sellerFeeBasisPoints", getU16Codec()],
      ["creators", getNullableCodec(getArrayCodec(CREATOR))],
      ["firstSaleSplit", getArrayCodec(SPLIT)],
    ]),
  ],
  [
    "update_pass",
    getStructCodec([
      ["name", getNullableCodec(getStringCodec())],
      ["symbol", getNullableCodec(getStringCodec())],
      ["uri", getNullableCodec(getStringCodec())],
      ["price", getNullableCodec(getU64Codec())],
      ["token", getNullableCodec(CURRENCY_CODEC)],
      ["supply", getNullableCodec(getU64Codec())],
      ["sellerFeeBasisPoints", getNullableCodec(getU16Codec())],
      ["creators", getNullableCodec(getNullableCodec(getArrayCodec(CREATOR)))],
      ["firstSaleSplit", getNullableCodec(getArrayCodec(SPLIT))],
    ]),
  ],
  ["mint_pass", getUnitCodec()],
]);

export type CodecInput<T> = T extends Codec<infer TFrom, any> ? TFrom : never;

export type CodecOutput<T> = T extends Codec<any, infer TTo> ? TTo : never;

export type PassInstructionDataInput = CodecInput<typeof PASS_INSTRUCTION_DATA>;
export type PassInstructionDataOutput = CodecOutput<
  typeof PASS_INSTRUCTION_DATA
>;
```