@lorisleiva 2024-03-15
@lithdew Almost identical traces on my side. I don't know if the results are biased by my machine's specs though. I also tried running this snippet on a code-sandbox and the TypeScript autocompletion was almost instant.

Are you able to compare the two branches on your side and tell me if you see a difference?