@lithdew 2024-03-15
Would you have the code sandbox available? I can try paste in parts of my project into it. Having a hard time linking to @solana/web3.js@master.