@lorisleiva 2024-03-15
Sure it's [here](https://codesandbox.io/p/devbox/codecs-performance-test-f85rs9) but the codesandbox only checks the current version. You'd need to check the branch locally to test the changes of this PR.