@lithdew 2024-03-17
Before:

```
Files:                         387
Lines of Library:            40626
Lines of Definitions:        92803
Lines of TypeScript:            81
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                153002
Symbols:                    163279
Types:                       82738
Instantiations:             642518
Memory used:               288025K
Assignability cache size:    42094
Identity cache size:           382
Subtype cache size:            161
Strict subtype cache size:     358
I/O Read time:               0.02s
Parse time:                  0.32s
ResolveModule time:          0.04s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.01s
Program time:                0.45s
Bind time:                   0.20s
Check time:                  1.88s
transformTime time:          0.01s
commentTime time:            0.00s
I/O Write time:              0.00s
printTime time:              0.01s
Emit time:                   0.01s
Total time:                  2.54s
```

After:
```
Files:                         388
Lines of Library:            40626
Lines of Definitions:        92819
Lines of TypeScript:            81
Lines of JavaScript:             0
Lines of JSON:                   0
Lines of Other:                  0
Identifiers:                153028
Symbols:                    163768
Types:                       82823
Instantiations:             400640
Memory used:               286162K
Assignability cache size:    42101
Identity cache size:           382
Subtype cache size:            161
Strict subtype cache size:     354
I/O Read time:               0.02s
Parse time:                  0.32s
ResolveModule time:          0.05s
ResolveTypeReference time:   0.01s
ResolveLibrary time:         0.01s
Program time:                0.44s
Bind time:                   0.18s
Check time:                  1.44s
transformTime time:          0.00s
commentTime time:            0.00s
I/O Write time:              0.00s
printTime time:              0.01s
Emit time:                   0.01s
Total time:                  2.08s
```

There is a noticeable improvement of 642518 -> 400640 instantiations processed. Wonder if there are other places this trick could be used too.